requirements.txt
streamlit
requests
BeautifulSoup
panda 